refactor: Long 타입의 Id를 Int 타입으로 변경

+++ app/src/main/java/woowacourse/shopping/data/dao/CartProductDao.kt
@@ -9,14 +9,14 @@ interface CartProductDao {
     fun getTotalCount(): Int
 
     @Query("SELECT quantity FROM cart_product WHERE product_id = :productId")
-    fun getQuantityByProductId(productId: Long): Int?
+    fun getQuantityByProductId(productId: Int): Int?
 
     @Query("SELECT COALESCE(SUM(quantity), 0) FROM cart_product")
     fun getTotalQuantity(): Int
 
     @Query("UPDATE cart_product SET quantity = :quantity WHERE product_id = :productId")
     fun updateQuantity(
-        productId: Long,
+        productId: Int,
         quantity: Int,
     )
 }
