+++ app/src/main/java/woowacourse/shopping/data/entity/RecentProductEntity.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "recent_product")
 data class RecentProductEntity(
-    @PrimaryKey @ColumnInfo(name = "product_id") val productId: Long,
+    @PrimaryKey @ColumnInfo(name = "product_id") val productId: Int,
     @ColumnInfo(name = "viewed_at") val viewedAt: Long = System.currentTimeMillis(),
 )
